+++ base-ui/prisma/schema.prisma
@@ -54,7 +54,7 @@ model Session {
 model User {
   id            String    @id @default(cuid())
   name          String?
-  email         String?   @unique
+  email         String?
   emailVerified DateTime? @map("email_verified")
   image         String?
   publicAddress String?   @map("public_address")
